+++ src/pages/UserLookbookCreate/ui/Edit/ColorPalette/ColorPalette.style.ts
