+++ src/pages/UserLookbookCreate/ui/Edit/ColorPalette/ColorPalette.tsx
