+++ src/pages/UserLookbookCreate/ui/Edit/Showcase/components/ClothesSlider/ClothesSlider.tsx
