+++ src/pages/UserLookbookCreate/ui/LookbookCreateHeader/LookbookCreateHeader.style.ts
